+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- `extension-crate-events.rb` Added buffering flushing before shutdown
 
 ## 0.1.0 - 2016-07-01
 ### Added